feat(ops): finalize docker-compose stack, observability, and config alignment
- Move project out of OneDrive to prevent sync conflicts and enable Docker workflows
- Provision PostgreSQL, Prometheus, and Grafana via docker-compose with persistent volumes
- Integrate Micrometer metrics across PulseEngine and StalkSchedulerService
- Configure Duration-native scheduler properties with SpEL binding; remove invalid @Min/@Max validators on Duration fields
- Register MonitoringProperties via @EnableConfigurationProperties for proper bean injection
- Resolve JPA/Hibernate schema validation: fix @Builder.Default, remove scale from double fields, align DECIMAL to DOUBLE PRECISION
- Set spring.jpa.hibernate.ddl-auto=none to defer schema management to Flyway
- Expose /actuator/prometheus endpoint and align with Spring Boot 4.x conventions

Establishes production-ready infrastructure, observability pipeline, and clean migration workflow.

+++ src/main/resources/db/migration/V1__create_stalks_and_pulses.sql
@@ -16,7 +16,7 @@ CREATE TABLE stalks (
                             CHECK (current_state IN ('HEALTHY', 'STRESSED', 'DEGRADED', 'DORMANT')),
 
     -- Cached Health Metrics (avoids expensive real-time calculations)
-                        health_index DECIMAL(5,2) DEFAULT 0.00,  -- 0-100 sliding window score
+                        health_index DOUBLE PRECISION DEFAULT 0.00,  -- 0-100 sliding window score
                         last_10_success_count INTEGER DEFAULT 0,
                         last_10_avg_latency_ms BIGINT DEFAULT 0,
                         consecutive_failures INTEGER DEFAULT 0,
